Updates for docker based ARM build
configure_linux_arm.mk and rules.linux.mk are updated to make
use of the environment variables provided by the docker file
(https://github.com/eclipse/openj9/pull/622).

This allows the names of the toolchain executables to be passed
from the docker file and also avoids the need to hard-code the
location of the global include files.

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>