Debug for unblock_spinlock_threads
Added debug to see which threads get notified during monitor_exit and when wait()
releases the monitor. Added debug to check for threads not being notified because
of the maxWakeThreads setting.

Noticed that we miss debug output from functions that were running when stopIprofilerThread
enables debug - because the state was privatized at the beginning of each function.
Stripped that out and marked jbkDebug volatile.